Fix viewport width.  Derive outdated versions from max agent version.
