+++ Server/Pages/Shared/_Layout.cshtml
@@ -132,7 +132,7 @@
     </footer>
 
 
-
+    <script src="~/site.js"></script>
     @RenderSection("Scripts", required: false)
 </body>
 </html>
